style: flex function now takes an additional argument that represents the value for flex-direction
The flex function now takes 3 arguments, align, justify, direction. align used to be direction, but I needed the direction name to represent the flexbox-direction value (because it made more sense). I still have to manually add the prefix -webkit-box-orient for internet explorer 10 users because it takes the values horizontal and vertical, instead of the usual row and column.